backend/requirements.txt
fastapi==0.116.1
firebase_admin==6.9.0
imagekitio==4.1.0
motor==3.3.2
pydantic==2.11.7
pymongo==4.6.0
python-dotenv==1.1.1
razorpay==1.4.2
uvicorn[standard]
pydantic[email]